refactor: move filter to standalone hooks
